Type change to byte
Changed types to byte where possible to save program space.

+++ LEDController/LEDController.h
@@ -23,7 +23,7 @@
 #define DELAY_TIME 100
 #define MAXBRIGHT 100
 
-extern int BRIGHTNESS;
+extern byte BRIGHTNESS;
 
 
 // initializs pins and values
@@ -53,10 +53,10 @@ extern void setSlider2();
 extern void setSlider3();
 
 //Global value to hold current value for each slider and for touchpad
-extern int button1;
-extern int button2;
-extern int button3;
-extern int button4;
+extern byte button1;
+extern byte button2;
+extern byte button3;
+extern byte button4;
 // used for calculating capactive touchpad
 extern unsigned int keys;
 extern unsigned int keys2;
@@ -67,9 +67,9 @@ extern CRGB leds[NUM_LEDS];
 
 // Tracks state of the touchpad to know when off or when on
 struct Touchkeys {
-  int buttonstate[8];
-  int ledState[8];
-  int LEDNum[8];
+  byte buttonstate[8];
+  byte ledState[8];
+  byte LEDNum[8];
   unsigned long old_time[8];
   
 };
@@ -79,9 +79,9 @@ extern Touchkeys touchkeys;
 
 
 //Used for the rotary encoder and to keep track of the current value (counter) 
-extern int counter;
-extern int switchCounter;
-extern int blinkCounter;
+extern byte counter;
+extern byte switchCounter;
+extern byte blinkCounter;
 
 
 #endif
